Refactored Functions into separate .js file
