+++ public/index.html
@@ -973,6 +973,7 @@
             </div>
         </div>
     </body>
+
     <!--Firebase-->
     <!-- Firebase App is always required and must be first -->
     <script src="https://www.gstatic.com/firebasejs/5.8.2/firebase-app.js"></script>
@@ -1000,6 +1001,7 @@
 
     <!-- Call JS File after DOM loads -->
     <script src="./index.js"></script>
+    <script src="./functions.js"></script>
     <script>
         if ("serviceWorker" in navigator) {
             navigator.serviceWorker.register("./sw.js").then(function() {
