+++ public/sw.js
@@ -9,7 +9,8 @@ self.addEventListener("install", function(e) {
                 "/Chart.js",
                 "/help.html",
                 "/about.html",
-                "./fontawesome-free-5.4.1-web/svgs/solid/chevron-circle-up.svg"
+                "./fontawesome-free-5.4.1-web/svgs/solid/chevron-circle-up.svg",
+                "./functions.js"
             ]);
         })
     );
